Reconfigured app.js to include routers
