+++ my-react-app/src/user-profile.css
@@ -1,5 +1,16 @@
 /*Emma CSS for the USER PROFILE PAGE...*/
-.user-profile{
+body{
     background-color: #6a012b; 
+    color: black; 
+    display: flex;
+    justify-content: center;
+}
+
+.userProf{
+    font-size: 4rem; 
     color: white; 
+    display: flex; 
+    margin-top: 5rem;
+    justify-content: center; 
+
 }
