--- my-react-app/src/user-profile.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="user-profile.css">
-    <title>Document</title>
-</head>
-<body>
-    <h1>User Profile</h1> 
-</body>
-</html>
